backend/src/requirements.txt
chromadb
ollama
django
django-ninja
django-ninja-extra
psycopg2-binary
django-ninja-jwt==5.3.7
django-cors-headers
pydantic==2.11.9
pydantic[email]
requests
gunicorn
whitenoise
celery
redis
django-redis
django-celery-beat
pywebpush
pytz
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-core
langgraph

# Dependencies for files app
pillow
pytesseract
yt-dlp
reportlab
cryptography